structure of the code and test was chenged
